+++ packages/core/src/odata/common/uri-conversion/get-filter.ts
@@ -1,5 +1,6 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
+import { isMoment } from 'moment';
 import { EntityBase } from '../entity';
 import { toStaticPropertyFormat } from '../../../util';
 import {
@@ -19,7 +20,6 @@ import {
 } from '../selectable';
 import { UriConverter } from '../request';
 import { isFilterLambdaExpression } from '../filter/filter-lambda-expression';
-import { convertToUriForEdmString } from './uri-value-converter';
 
 // eslint-disable-next-line valid-jsdoc
 /**
@@ -196,10 +196,13 @@ export function createGetFilter(uriConverter: UriConverter) {
       return param.toString();
     }
     if (typeof param === 'string') {
-      return convertToUriForEdmString(param);
+      return uriConverter.convertToUriFormat(param, 'Edm.String');
     }
     if (param instanceof FilterFunction) {
       return filterFunctionToString(param, parentFieldNames);
+    }
+    if (isMoment(param)) {
+      return uriConverter.convertToUriFormat(param, 'Edm.DateTimeOffset');
     }
     return [...parentFieldNames, param._fieldName].join('/');
   }