+++ packages/generator/test/generator.spec.ts
@@ -1,11 +1,31 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
+
 import { ODataVersion } from '@sap-cloud-sdk/util';
 import { SourceFile, ClassDeclaration, FunctionDeclaration } from 'ts-morph';
-import { generateProject } from '../src';
+import { generateProject, GeneratorOptions } from '../src';
 import { createOptions } from './test-util/create-generator-options';
 
 describe('generator', () => {
   let files: SourceFile[];
+
+  describe('common', () => {
+    it('copies the additional files matching the glob.', async () => {
+      const project = await generateProject(
+        createOptions({
+          inputDir: '../../test-resources/service-specs/v2/API_TEST_SRV',
+          additionalFiles: '../../test-resources/*.md'
+        })
+      );
+
+      const sourceFiles = project!.getSourceFiles();
+      expect(
+        sourceFiles.find(file => file.getBaseName() === 'some-test-markdown.md')
+      ).toBeDefined();
+      expect(
+        sourceFiles.find(file => file.getBaseName() === 'CHANGELOG.md')
+      ).toBeDefined();
+    });
+  });
 
   describe('v2', () => {
     beforeAll(async () => {
@@ -101,7 +121,8 @@ function checkStaticProperties(entityClass: ClassDeclaration): void {
 }
 
 async function getGeneratedFiles(
-  oDataVersion: ODataVersion
+  oDataVersion: ODataVersion,
+  generatorOptionsX?: GeneratorOptions
 ): Promise<SourceFile[]> {
   const project = await generateProject(
     createOptions({